Split activity/fragment extensions to app and support
