+++ kotlinx-android-app/build.gradle
@@ -23,5 +23,5 @@ android {
 }
 
 dependencies {
-    implementation "com.android.support:support-v4:$support_version"
+
 }
