+++ kotlinx-android-support-v4/src/main/AndroidManifest.xml
@@ -1,0 +1,1 @@
+<manifest package="kotlinx.android.support.v4" />
