+++ kotlinx-android-support-v4/src/main/kotlin/kotlinx/android/support/v4/FragmentExt.kt
@@ -1,4 +1,4 @@
-package kotlinx.android.activity
+package kotlinx.android.support.v4
 
 import android.support.annotation.ColorRes
 import android.support.annotation.StringRes
@@ -11,15 +11,7 @@ fun Fragment.toast(@StringRes resourceId: Int, length: Int = Toast.LENGTH_SHORT)
     activity.toast(resourceId, length)
 }
 
-fun android.app.Fragment.toast(@StringRes resourceId: Int, length: Int = Toast.LENGTH_SHORT) {
-    activity.toast(resourceId, length)
-}
-
 fun Fragment.toast(message: String, length: Int = Toast.LENGTH_SHORT) {
-    activity.toast(message, length)
-}
-
-fun android.app.Fragment.toast(message: String, length: Int = Toast.LENGTH_SHORT) {
     activity.toast(message, length)
 }
 
@@ -27,17 +19,7 @@ fun Fragment.colorFromResource(@ColorRes colorResource: Int): Int {
     return ContextCompat.getColor(activity, colorResource)
 }
 
-fun android.app.Fragment.colorFromResource(@ColorRes colorResource: Int): Int {
-    return ContextCompat.getColor(activity, colorResource)
-}
-
 inline fun Fragment.fragmentTransaction(function: FragmentTransaction.() -> FragmentTransaction) {
-    fragmentManager.beginTransaction()
-            .function()
-            .commit()
-}
-
-inline fun android.app.Fragment.fragmentTransaction(function: android.app.FragmentTransaction.() -> android.app.FragmentTransaction) {
     fragmentManager.beginTransaction()
             .function()
             .commit()
